+++ src/Command/AppConfigGetCommand.php
@@ -5,10 +5,6 @@
 
 namespace DevopsToolAppOrchestration\Command;
 
-use DevopsToolCore\MonologConsoleHandlerAwareTrait;
-use Exception;
-use Psr\Log\LoggerInterface;
-use Psr\Log\NullLogger;
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -17,29 +13,6 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class AppConfigGetCommand extends AbstractCommand
 {
-    use MonologConsoleHandlerAwareTrait;
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
-
-    /**
-     * AppListCommand constructor.
-     *
-     * @param LoggerInterface $logger
-     * @param null            $name
-     */
-    public function __construct(
-        LoggerInterface $logger = null,
-        $name = null
-    ) {
-        if (is_null($logger)) {
-            $logger = new NullLogger();
-        }
-        $this->logger = $logger;
-        parent::__construct($name);
-    }
-
     protected function configure()
     {
         $this->setName('app:config:get')
@@ -50,35 +23,52 @@ class AppConfigGetCommand extends AbstractCommand
             ->addArgument(
                 'key',
                 InputArgument::REQUIRED,
-                'Configuration key to grab value for. Multiple levels should be separated by period (path.to.key).'
+                'Configuration key to grab value for. Multiple levels should be separated by a forward slash (path/to/key).'
             )
-            ->addOption('app', null, InputOption::VALUE_REQUIRED, 'App id to get configuration data about.');
+            ->addOption(
+                'app',
+                null,
+                InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY,
+                'Application code to get configuration data about.'
+            )
+            ->addOption('all', null, InputOption::VALUE_NONE, 'Show configuration for all applications.');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $this->injectOutputIntoLogger($output, $this->logger);
-        $applicationCode = $input->getOption('app');
-        $application = $this->config['applications'][$applicationCode];
-
+        $applications = $this->getApplications($input);
         $key = $input->getArgument('key');
-        $value = $application;
-        $parts = explode('.', $key);
-        foreach ($parts as $part) {
-            if (!array_key_exists($part, $value)) {
-                throw new Exception(
-                    "Key \"$key\" not found in application $applicationCode (${application['name']}) configuration."
-                );
-            }
-            $value = $value[$part];
-        }
+        // Remove "/*" suffix when searching by array key
+        $searchKey = preg_replace('%/\*$%', '', $key);
 
         $outputTable = new Table($output);
         $outputTable
-            ->setHeaders(['Key', 'Value'])
-            ->addRow([$key, $value])
-            ->render();
+            ->setHeaders(['Application Code', 'Key', 'Value']);
 
+        foreach ($applications as $code => $application) {
+            $value = $application->getArrayCopy();
+            $parts = explode('/', $searchKey);
+            foreach ($parts as $part) {
+                if (is_array($value) && array_key_exists($part, $value)) {
+                    $value = $value[$part];
+                } else {
+                    $value = null;
+                    break;
+                }
+            }
+
+            if (is_null($value)) {
+                $displayValue = 'NULL';
+            } elseif (is_array($value)) {
+                $displayValue = print_r($value, true);
+            } else {
+                $displayValue = $value;
+            }
+
+            $outputTable->addRow([$code, $key, $displayValue]);
+        }
+
+        $outputTable->render();
         return 0;
     }