+++ CHANGELOG.md
@@ -2,6 +2,7 @@ DevOps Tool: Application Orchestration
 ==============================================
 
 # 0.2.0 (Unreleased)
+- Refactored app:destroy, app:list, app:config:list, and app:config:get
 - Refactored application configuration to pull from local config rather than app setup repo
 
 # 0.1.0